fix register step
